requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
fastapi==0.103.2
greenlet==3.0.0
h11==0.14.0
idna==3.4
MarkupSafe==2.1.3
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
setuptools==68.0.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
Werkzeug==3.0.0
wheel==0.37.1
